Alpha 1.3.1 - Minor Bug Fixes
Star background in main menu now spins with the galaxy
Changed the player jump sound
Fixed the checkpoint being able to be activated more than once during phase 2
Updated most prefabs to default at coordinates (0,0)
Changed hitbox size of escape spaceship for phase 2
Walls-tilemap has been properly place in the 'terrain' sorting layer (instead of default)
Fixed a bug where pressing ESC during the game over screen would freeze the time scale

+++ 226/Assets/Scripts/Objects/Object_Spinning.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Object_Spinning : MonoBehaviour
 {
 
-    public int rotationSpeed;
+    public float rotationSpeed;
 
     // Rotate the object every frame
     void Update()
